#9 Starting work on the channel communication layer
The basic idea is an object that takes the events from the client and
can decode the json and send it to the proper channels, i.e. the player
channel, the physics channel, or the player creation channel.
Commenting a lot of code out to implement later once the communication
layer.