#15 Implemented in a limited capicity
the outline of sending all the events in a single message works, but
there is a question of how to order the messages properly as we need to
ensure that objects are added before an update is called on them.
Current convention is to read the array in order but that is a weak
convention to follow in my opinion.